+++ dotnet_src/Math/LinearAlgebra/MatrixDimensions.cs
@@ -2,8 +2,8 @@ namespace LinearAlgebra;
 
 public record MatrixDimensions
 {
-    public required uint Rows { get; init; }
-    public required uint Cols { get; init; }
+    public required ushort Rows { get; init; }
+    public required ushort Cols { get; init; }
 
     public override string ToString()
     {
